+++ TRABAJO_FINAL_GRUPO_STROKES/src/main/java/ar/edu/unju/fi/tpfinal/service/IEmployeeService.java
@@ -11,5 +11,5 @@ public interface IEmployeeService {
 	public void deleteEmployee(Long employeeNumber);
 	public Optional<Employee> getEmployee(Long employeeNumber);
 	public List<Employee> getEmployees();
-	public Employee searchEmployee(int numero, String email);
+	public Employee searchEmployee(Long numero, String email);
 }
